+++ dev-packages/browser-integration-tests/suites/sessions/disabled-on-load/init.js
@@ -1,0 +1,11 @@
+import * as Sentry from '@sentry/browser';
+
+window.Sentry = Sentry;
+
+Sentry.init({
+  dsn: 'https://public@dsn.ingest.sentry.io/1337',
+  release: '0.1',
+});
+
+// Disable right at the start
+Sentry.getClient().setSessionTrackingEnabled(false);